refactored codebase for GroupMessage
