+++ src/views/Chat/components/ChatView/components/GroupMessage/components/GroupMessageList/components/ImageMessage/index.js
