+++ src/views/Chat/components/ChatView/components/GroupMessage/components/GroupMessageList/components/LastCardIndicator/index.js
