+++ src/views/Chat/components/ChatView/components/GroupMessage/components/GroupMessageList/components/ReservationMessage/index.js
@@ -1,0 +1,1 @@
+export { default } from './ReservationMessage'
