+++ src/views/Chat/components/ChatView/components/GroupMessage/components/GroupMessageList/components/index.js
@@ -1,0 +1,5 @@
+export { default as ImageMessage } from './ImageMessage'
+export { default as LastCardIndicator } from './LastCardIndicator'
+export { default as FileMessage } from './FileMessage'
+export { default as TextMessage } from './TextMessage'
+export { default as ReservationMessage } from './ReservationMessage'
