--- src/views/Chat/components/ChatView/components/GroupMessage/components/GroupedMessages/index.js
@@ -1,1 +1,0 @@
-export { default } from './GroupedMessages'
