+++ src/views/Chat/components/ChatView/components/GroupMessage/components/index.js
@@ -1,2 +1,2 @@
 export { default as GroupBreak } from './GroupBreak'
-export { default as GroupedMessages } from './GroupedMessages'
+export { default as GroupMessageList } from './GroupMessageList'
