--- src/views/Chat/components/ChatView/components/MessageCard/components/index.js
@@ -1,3 +1,0 @@
-export { default as FileMessage } from './FileMessage'
-export { default as ImageMessage } from './ImageMessage'
-export { default as LastCardIndicator } from './LastCardIndicator'
