--- src/views/Chat/components/ChatView/components/MessageCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './MessageCard'
