--- src/views/Chat/components/ChatView/components/MessageReservation/index.js
@@ -1,1 +1,0 @@
-export { default } from './MessageReservation'
