+++ src/views/Chat/gql/subscription.js
@@ -1,42 +1,17 @@
 import gql from 'graphql-tag'
+import { ZALO_MESSAGE } from './fragment'
 
 const ON_ZALO_MESSAGE_RECEIVED = gql`
 	subscription {
 		onZaloMessageReceived {
-			id
-			type
-			from {
-				displayName
-				id
-				avatar
-			}
-			to {
-				displayName
-				id
-				avatar
-			}
-			timestamp
-			content
+			...ZaloMessage
 		}
 	}
+	${ZALO_MESSAGE}
 `
 const ON_ZALO_MESSAGE_CREATED = gql`
 	subscription onZaloMessageCreated($filter: OnZaloMessageCreatedInput!) {
 		onZaloMessageCreated(filter: $filter) {
-			id
-			from {
-				displayName
-				id
-				avatar
-			}
-			type
-			to {
-				displayName
-				id
-				avatar
-			}
-			timestamp
-			content
 			attachments {
 				payload {
 					thumbnail
@@ -46,48 +21,27 @@ const ON_ZALO_MESSAGE_CREATED = gql`
 				}
 				type
 			}
+			...ZaloMessage
 		}
 	}
+	${ZALO_MESSAGE}
 `
 const ON_ZALO_MESSAGE_RECEIVED_FROM_USER = gql`
 	subscription onZaloMessageReceived($filter: OnZaloMessageReceivedInput) {
 		onZaloMessageReceived(filter: $filter) {
-			id
-			from {
-				displayName
-				id
-				avatar
-			}
-			to {
-				displayName
-				id
-				avatar
-			}
-			timestamp
-			content
+			...ZaloMessage
 		}
 	}
+	${ZALO_MESSAGE}
 `
 
 const ON_ZALO_MESSAGE_CREATED_TO_USER = gql`
 	subscription onZaloMessageCreated($filter: OnZaloMessageCreatedInput) {
 		onZaloMessageCreated(filter: $filter) {
-			id
-			from {
-				displayName
-				id
-				avatar
-			}
-			type
-			to {
-				displayName
-				id
-				avatar
-			}
-			timestamp
-			content
+			...ZaloMessage
 		}
 	}
+	${ZALO_MESSAGE}
 `
 
 export {
